+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 **[0.9.4.1]**
 * Fix multi-create netpoint
+* Fix Link PatchPanel Model and Type
 
 **[0.9.4.0]**
 * First Commit